+++ reactComponent/example/src/main.tsx
@@ -1,7 +1,10 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import 'ia-front-ref-assistant/style.css'
+import { mountIaFrontRefAssistant } from 'ia-front-ref-assistant'
+import config from '../iafrontrefassistant.config'
 import App from './App'
+
+mountIaFrontRefAssistant(config)
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>